fixing admin roles, makign delete/edit buttons appear only for proper permissions

+++ web-ui/src/Entries/View.js
@@ -6,14 +6,14 @@ import React from 'react';
 /*attribution for emojis: https://medium.com/@seanmcp/%EF%B8%8F-how-to-use-emojis-in-react-d23bbf608bf7*/
 
 const Emoji = props => (
-    <span
-        className="emoji"
-        role="img"
-        aria-label={props.label ? props.label : ""}
-        aria-hidden={props.label ? "false" : "true"}
-    >
-        {props.symbol}
-    </span>
+	<span
+		className="emoji"
+		role="img"
+		aria-label={props.label ? props.label : ""}
+		aria-hidden={props.label ? "false" : "true"}
+	>
+		{props.symbol}
+	</span>
 );
 
 function Entry({ entry, company, session }) {
@@ -33,40 +33,48 @@ function Entry({ entry, company, session }) {
 		let yes = 0;
 		let no = 0;
 		if (entry.drug_test.toString() == "false") {
-			return (<p className="booleanFalse"> <span className="black-txt">Drug test?</span> This company does <b>not</b> drug test.<Emoji symbol="🤩" label="starstruck"/></p> );
+			return (<p className="booleanFalse"> <span className="black-txt">Drug test?</span> This company does <b>not</b> drug test.<Emoji symbol="🤩" label="starstruck" /></p>);
 		}
 		else {
-			return (<p className="booleanTrue"><span className="black-txt">Drug test? </span>This company drug tests. <Emoji symbol="😬" label="grimace"/></p>);
+			return (<p className="booleanTrue"><span className="black-txt">Drug test? </span>This company drug tests. <Emoji symbol="😬" label="grimace" /></p>);
 		}
-}
-
-function renderContinuous() {
-	if (entry.continuous_drug_testing.toString() == "false") {
-		return (<p> Continuous testing? This company does <b>not</b> drug test continuously. <Emoji symbol="🤩" label="starstruck"/></p>);
 	}
-	else {
-		return (<p> Continuous testing? This company does drug test continuously. <Emoji symbol="😬" label="grimace"/></p>);
-	}
-}
 
-return (
-	<Container className="company-details">
-		<Row>
-			<Col>
-				{/* <h5> Entry {entry.id} </h5> */}
-				{coop_cycle()}
-				{renderDrugTest()}
-				{renderContinuous()}
-				<br></br>
-				{additional_notes()}
-				<br />
-			</Col>
-			<Col>
-				<Button variant="primary-outline" onClick={() => delete_entry(entry.id)}>Delete Entry</Button>
-			</Col>
-		</Row>
-	</Container>
-);
+	function renderContinuous() {
+		if (entry.continuous_drug_testing.toString() == "false") {
+			return (<p> Continuous testing? This company does <b>not</b> drug test continuously. <Emoji symbol="🤩" label="starstruck" /></p>);
+		}
+		else {
+			return (<p> Continuous testing? This company does drug test continuously. <Emoji symbol="😬" label="grimace" /></p>);
+		}
+	}
+
+	function privs() {
+		if (session && ((session.user_id.toString() === entry.user_id.toString()) || (session.admin === true || session.admin === "true"))) {
+			return (
+				<Col>
+					<Button variant="primary-outline" onClick={() => delete_entry(entry.id)}>Delete Entry</Button>
+				</Col>
+			);
+		}
+	}
+
+	return (
+		<Container className="company-details">
+			<Row>
+				<Col>
+					{/* <h5> Entry {entry.id} </h5> */}
+					{coop_cycle()}
+					{renderDrugTest()}
+					{renderContinuous()}
+					<br></br>
+					{additional_notes()}
+					<br />
+				</Col>
+				{privs()}
+			</Row>
+		</Container>
+	);
 }
 
 function ViewEntries({ company, session }) {
@@ -78,8 +86,8 @@ function ViewEntries({ company, session }) {
 	return (
 		<Row>
 			<Col>
-			<br>
-			</br>
+				<br>
+				</br>
 				<h2> User Entries</h2>
 				{entries}
 			</Col>
